+++ src/ts/components/CControlAutocomplete.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="c-control c-control-autocomplete">
         <label class="c-control__label block font-medium mb-1 text-slate-800 text-sm"
-               v-bind:class="{ 'text-error': invalid, 'after:content-[\'*\'] after:text-red-400': required }"
+               v-bind:class="{ '!text-error': invalid, 'after:content-[\'*\'] after:text-red-400': required }"
                v-bind:for="name"
                v-if="label"
         >
@@ -12,6 +12,7 @@
              v-bind:class="{ '!border-b-slate-100' : disabled, '!border-b-error' : invalid }"
         >
             <input class="c-control__field c-control-autocomplete__field bg-transparent block disabled:bg-white disabled:cursor-not-allowed disabled:text-slate-400 focus:outline-0 h-8 leading-none peer px-1 read-only:pointer-events-none text-slate-800 text-sm w-full" autocomplete="off" type="text"
+                   v-bind:class="{ '!text-error' : invalid }"
                    v-bind:disabled="disabled"
                    v-bind:id="name"
                    v-bind:placeholder="placeholder"
@@ -29,6 +30,8 @@
                  v-bind:src="avatar"
                  v-if="avatarField && avatar"
             />
+
+            <c-icon class="cursor-pointer h-4 stroke-slate-400 !stroke-2 w-4" icon="x" v-on:click="reset" />
 
             <c-icon class="h-6 peer-focus:stroke-highlight stroke-slate-800 w-6"
                     v-bind:class="{
@@ -71,7 +74,7 @@
         </div>
 
         <div class="c-control__notes c-control-autocomplete__notes text-slate-800 text-sm"
-             v-bind:class="{ 'text-error': invalid }"
+             v-bind:class="{ '!text-error': invalid }"
              v-if="$slots.notes"
         >
             <slot name="notes"></slot>
@@ -194,6 +197,15 @@
         }
 
         return obj[prop];
+    }
+
+    function reset() {
+        searchString.value = '';
+        searchResults.value = [];
+        selectedValue.value = null;
+        avatar.value = '';
+
+        emits('update:modelValue', null);
     }
 
     function search(searchValue: any) {