+++ src/ts/components/CControlNumber.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="c-control c-control-number">
         <label class="c-control__label block font-medium mb-1 text-slate-800 text-sm"
-               v-bind:class="{ 'text-error': invalid, 'after:content-[\'*\'] after:text-red-400': required }"
+               v-bind:class="{ '!text-error': invalid, 'after:content-[\'*\'] after:text-red-400': required }"
                v-bind:for="name"
                v-if="label"
         >
@@ -12,6 +12,7 @@
              v-bind:class="{ '!border-b-slate-100' : disabled, '!border-b-error' : invalid }"
         >
             <input class="c-control__field c-control-text__field bg-transparent block disabled:bg-white disabled:cursor-not-allowed disabled:text-slate-400 focus:outline-0 h-8 leading-none peer px-1 read-only:pointer-events-none text-slate-800 text-sm w-full" type="number"
+                   v-bind:class="{ '!text-error': invalid }"
                    v-bind:disabled="disabled"
                    v-bind:id="name"
                    v-bind:max="max"
@@ -39,7 +40,7 @@
         </div>
 
         <div class="c-control__notes c-control-number__notes text-slate-800 text-sm"
-             v-bind:class="{ 'text-error': invalid }"
+             v-bind:class="{ '!text-error': invalid }"
              v-if="$slots.notes"
         >
             <slot name="notes"></slot>