+++ CHANGELOG.md
@@ -13,3 +13,8 @@ All notable changes to `@teamfurther/cinderblock` will be documented in this fil
 
 ## 1.0.3 - 2023-05-25
 - Updated JS dependencies.
+
+## 1.0.4 - 2023-07-25
+- Adjusted invalid state CSS.
+- Added reset method to CControlAutocomplete.
+- Updated JS dependencies.